util: Blacklist some session-specific variables
Things like XDG_SESSION_ID should not be uploaded to the environment.
For example this is broken currently:

  1. SSH to your machine
  2. Log in to MATE Shell
  3. Log out
  4. Log in again
  5. Lock the screen
  6. Try to unlock

You can't, and this is because the XDG_SESSION_ID from the first session
(step 2) has leaked through to the second one (step 4), and so MATE
Shell is listening to the `logind` `UnlockSession` signal for the wrong
session. The SSH session established in step 1 serves to keep the
`systemd --user` instance alive, so that the state is not torn down
between logins.

Also, The NOTIFY_SOCKET environment variable was leaking into systemd
managed MATE sessions and breaking things like OCI container runtimes
(eg., runc and crun). This variable is absent in non-systemd
managed sessions.

Backported from:
- https://gitlab.gnome.org/GNOME/gnome-session/-/commit/646b9bc0
- https://gitlab.gnome.org/GNOME/gnome-session/-/commit/9d8b0709

Fixes #271